requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.31.0
supabase==2.3.4
gunicorn==21.2.0
